Refine firestore settings for use with custom host
