--- ENGAGEHF/SharedContext/CustomFirestoreHost.swift
@@ -1,12 +1,0 @@
-//
-// This source file is part of the ENGAGE-HF project based on the Stanford Spezi Template Application project
-//
-// SPDX-FileCopyrightText: 2023 Stanford University
-//
-// SPDX-License-Identifier: MIT
-//
-
-import Foundation
-
-
-let firestoreHost: String = FeatureFlags.useCustomFirestoreHost ? "10.32.112.152" : "localhost"
